rename bme280 to bme280_i2c
